Reduce CPU usage of CSS animation
This special override will stop the background animation and only play it through once. As this is a page visitors might leave open in the background, it's very energy wasteful to keep animating. By stopping the CSS after a bit (30sec) the CPU usage on this page drops from 20-30% down to ~0.3%.